修复PostgreSQL数据类型兼容性问题
- 将所有实体中的 tinyint 类型改为 smallint
- 修复 User.gender, User.status, Product.status 等字段
- 修复 Order.orderStatus, Order.payStatus 等字段
- 修复 Merchant.status, Category.level, Category.status 等字段
- 修复 Cart.selected, MerchantProduct.status 等字段
- 修复 MerchantRecharge.status, MerchantWithdrawalInfo 等字段
- 修复 FundOperation.operationType, MerchantCreditRating 等字段
- 修复 MerchantWithdrawal.status 字段

现在所有实体都与PostgreSQL兼容

+++ ecommerce-backend/src/modules/cart/entities/cart.entity.ts
@@ -26,7 +26,7 @@ export class Cart {
   @Column({ type: 'int', default: 1 })
   quantity: number;
 
-  @Column({ type: 'tinyint', default: 1, comment: '是否选中 1是 0否' })
+  @Column({ type: 'smallint', default: 1, comment: '是否选中 1是 0否' })
   selected: number;
 
   @CreateDateColumn({ name: 'create_time', type: 'datetime' })
